refactor(consumption): extract pollution stripper into sibling helper (Refs #563 phase: pump_display_parser)
Move stripPumpDisplayPollution and its three private regexes into
_pump_display_pollution.dart so pump_display_parser.dart drops back
under the 300-LOC budget. Re-exported from the parser file so the
existing test imports continue to resolve unchanged.

LOC: 335 -> 269 (parser); 79 (new sibling helper).
